partner_chains_dolos_data_sources::client::api

Trait MiniBFApi

pub trait MiniBFApi {

    // Required methods
    fn addresses_utxos<'life0, 'async_trait>(
        &'life0 self,
        address: MainchainAddress,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AddressUtxoContentInner>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn addresses_transactions<'life0, 'async_trait>(
        &'life0 self,
        address: MainchainAddress,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AddressTransactionsContentInner>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn assets_transactions<'life0, 'async_trait>(
        &'life0 self,
        asset_id: AssetId,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AssetTransactionsInner>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn assets_addresses<'life0, 'async_trait>(
        &'life0 self,
        asset_id: AssetId,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AssetAddressesInner>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocks_latest<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<BlockContent, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocks_by_id<'life0, 'async_trait>(
        &'life0 self,
        id: impl 'async_trait + Into<McBlockId> + Send,
    ) -> Pin<Box<dyn Future<Output = Result<BlockContent, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocks_slot<'life0, 'async_trait>(
        &'life0 self,
        slot_number: McSlotNumber,
    ) -> Pin<Box<dyn Future<Output = Result<BlockContent, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocks_next<'life0, 'async_trait>(
        &'life0 self,
        hash: impl 'async_trait + Into<McBlockId> + Send,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<BlockContent>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocks_txs<'life0, 'async_trait>(
        &'life0 self,
        id: impl 'async_trait + Into<McBlockId> + Send,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn epochs_blocks<'life0, 'async_trait>(
        &'life0 self,
        epoch_number: McEpochNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn epochs_parameters<'life0, 'async_trait>(
        &'life0 self,
        epoch_number: McEpochNumber,
    ) -> Pin<Box<dyn Future<Output = Result<EpochParamContent, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn epochs_stakes_by_pool<'life0, 'async_trait>(
        &'life0 self,
        epoch_number: McEpochNumber,
        pool_id: impl 'async_trait + Into<McPoolId> + Send,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<EpochStakePoolContentInner>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pools_history<'life0, 'async_trait>(
        &'life0 self,
        pool_id: impl 'async_trait + Into<McPoolId> + Send,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PoolHistoryInner>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pools_extended<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PoolListExtendedInner>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn scripts_datum_hash<'life0, 'life1, 'async_trait>(
        &'life0 self,
        datum_hash: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Value>, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn transaction_by_hash<'life0, 'async_trait>(
        &'life0 self,
        tx_hash: McTxHash,
    ) -> Pin<Box<dyn Future<Output = Result<TxContent, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transactions_utxos<'life0, 'async_trait>(
        &'life0 self,
        tx_hash: McTxHash,
    ) -> Pin<Box<dyn Future<Output = Result<TxContentUtxo, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn genesis<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<GenesisContent, DataSourceError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Mini Blockfrost API interface

Required Methods

fn addresses_utxos<'life0, 'async_trait>( &'life0 self, address: MainchainAddress, ) -> Pin<Box<dyn Future<Output = Result<Vec<AddressUtxoContentInner>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

UTXOs of the address.

fn addresses_transactions<'life0, 'async_trait>( &'life0 self, address: MainchainAddress, ) -> Pin<Box<dyn Future<Output = Result<Vec<AddressTransactionsContentInner>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Transactions on the address.

fn assets_transactions<'life0, 'async_trait>( &'life0 self, asset_id: AssetId, ) -> Pin<Box<dyn Future<Output = Result<Vec<AssetTransactionsInner>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List of specific asset transactions.

fn assets_addresses<'life0, 'async_trait>( &'life0 self, asset_id: AssetId, ) -> Pin<Box<dyn Future<Output = Result<Vec<AssetAddressesInner>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List of addresses containing a specific asset.

fn blocks_latest<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<BlockContent, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the latest block available to the backends, also known as the tip of the blockchain.

fn blocks_by_id<'life0, 'async_trait>( &'life0 self, id: impl 'async_trait + Into<McBlockId> + Send, ) -> Pin<Box<dyn Future<Output = Result<BlockContent, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the content of a requested block.

fn blocks_slot<'life0, 'async_trait>( &'life0 self, slot_number: McSlotNumber, ) -> Pin<Box<dyn Future<Output = Result<BlockContent, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the content of a requested block for a specific slot.

fn blocks_next<'life0, 'async_trait>( &'life0 self, hash: impl 'async_trait + Into<McBlockId> + Send, ) -> Pin<Box<dyn Future<Output = Result<Vec<BlockContent>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of blocks following a specific block.

fn blocks_txs<'life0, 'async_trait>( &'life0 self, id: impl 'async_trait + Into<McBlockId> + Send, ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the transactions within the block.

fn epochs_blocks<'life0, 'async_trait>( &'life0 self, epoch_number: McEpochNumber, ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the blocks minted for the epoch specified.

fn epochs_parameters<'life0, 'async_trait>( &'life0 self, epoch_number: McEpochNumber, ) -> Pin<Box<dyn Future<Output = Result<EpochParamContent, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the protocol parameters for the epoch specified.

fn epochs_stakes_by_pool<'life0, 'async_trait>( &'life0 self, epoch_number: McEpochNumber, pool_id: impl 'async_trait + Into<McPoolId> + Send, ) -> Pin<Box<dyn Future<Output = Result<Vec<EpochStakePoolContentInner>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the active stake distribution for the epoch specified by stake pool.

fn pools_history<'life0, 'async_trait>( &'life0 self, pool_id: impl 'async_trait + Into<McPoolId> + Send, ) -> Pin<Box<dyn Future<Output = Result<Vec<PoolHistoryInner>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

History of stake pool parameters over epochs.

fn pools_extended<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Vec<PoolListExtendedInner>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List of registered stake pools with additional information.

fn scripts_datum_hash<'life0, 'life1, 'async_trait>( &'life0 self, datum_hash: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<Vec<Value>, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query JSON value of a datum by its hash.

fn transaction_by_hash<'life0, 'async_trait>( &'life0 self, tx_hash: McTxHash, ) -> Pin<Box<dyn Future<Output = Result<TxContent, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return content of the requested transaction.

fn transactions_utxos<'life0, 'async_trait>( &'life0 self, tx_hash: McTxHash, ) -> Pin<Box<dyn Future<Output = Result<TxContentUtxo, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the inputs and UTXOs of the specific transaction.

fn genesis<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<GenesisContent, DataSourceError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the information about blockchain genesis.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.